Avoid time flooring to millis.
Going via milliseconds truncates the nano component of Instant. This leads to vairous weird results when trying to work out remaining time.
The use of millis was done as an optimisation, since getting the nano time is slightly more expensive, but in benchmarks the difference seems to be insignificant/noise.